Refactoring part 1 (not ready)
