--- src/test/java/planner/AppTest.java
@@ -1,13 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package planner;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class AppTest {
-    @Test public void testAppHasAGreeting() {
-        assertNotNull("app should have a greeting", App.getGREETING());
-    }
-}
